+++ components/Admin/Session/Dashboard/Form/EditSessionForm/index.tsx
@@ -1,156 +1,23 @@
-import { FC, useCallback, useEffect, useState } from 'react'
+import { FC } from 'react'
 import {
   TextInput,
   BooleanInput,
   SimpleFormProps,
-  useInput,
-  InputProps,
   required,
-  TextFieldProps,
   Labeled,
-  Identifier,
   RadioButtonGroupInput,
   FormDataConsumer,
   FormWithRedirect,
   SaveButton,
   DeleteButton
 } from 'react-admin'
-import { Color, ColorPicker } from 'material-ui-color'
-import {
-  TextField as TextInputMU,
-  Grid,
-  makeStyles,
-  Toolbar,
-  Box,
-  Button
-} from '@material-ui/core'
-import { Add as AddIcon } from '@material-ui/icons'
+import { Grid, Toolbar, Box } from '@material-ui/core'
 import isColor from 'is-color'
-import { useForm, useField } from 'react-final-form'
-import { Storage } from 'aws-amplify'
-import { DropzoneDialog } from 'material-ui-dropzone'
 import SimplePreview from './SimplePreview'
+import { ColorInput, ImageInput } from '../../../parts'
 
 const colorValidator = (color: string) => {
   return isColor(color) ? null : '入力内容が間違っています'
-}
-
-const ColorInput: FC<InputProps<TextFieldProps>> = (props) => {
-  const {
-    input: { name, onChange, value, ...rest },
-    meta: { touched, error },
-    isRequired
-  } = useInput(props)
-  const { change } = useForm()
-  const handleChange = useCallback(
-    (color: Color) => {
-      change(name, `#${color.hex}`)
-    },
-    [change]
-  )
-
-  return (
-    <Grid container alignItems="center">
-      <Grid item xs={1}>
-        <ColorPicker
-          value={value}
-          hideTextfield
-          disableAlpha
-          onChange={handleChange}
-        />
-      </Grid>
-      <Grid item xs={11}>
-        <TextInputMU
-          name={name}
-          label={props.label}
-          onChange={onChange}
-          error={!!(touched && error)}
-          helperText={touched && error ? error : ' '}
-          required={isRequired}
-          value={value}
-          {...rest}
-        />
-      </Grid>
-    </Grid>
-  )
-}
-
-const useStyles = makeStyles(() => ({
-  image: {
-    height: 100,
-    width: 'initial',
-    maxWidth: '100%'
-  }
-}))
-
-const ImageInput: FC<InputProps<TextFieldProps>> = (props) => {
-  const classes = useStyles()
-  const { source, label } = props
-  const {
-    input: { value }
-  } = useField(source)
-  const {
-    input: { value: sessionId }
-  } = useField<Identifier>('id')
-  const { change } = useForm()
-  const [src, setSrc] = useState('')
-  const [open, setOpen] = useState(false)
-  const handleClose = useCallback(() => setOpen(false), [])
-  const handleOpen = useCallback(() => setOpen(true), [])
-  const handleSave = useCallback(
-    async ([file]: File[]) => {
-      const res = await Storage.put(`${sessionId}/${file.name}`, file, {
-        level: 'public'
-      })
-      change(source, (res as { key: string }).key)
-      handleClose()
-    },
-    [change, handleClose, sessionId]
-  )
-
-  useEffect(() => {
-    if (!value) return
-    Storage.get(value, { level: 'public' }).then(
-      (val) => typeof val === 'string' && setSrc(val)
-    )
-  }, [value])
-
-  return (
-    <Labeled label={label} fullWidth>
-      <>
-        <Grid item xs={5}>
-          {src && <img src={src} className={classes.image} />}
-          <Button onClick={handleOpen}>
-            {src ? (
-              '変更'
-            ) : (
-              <>
-                <AddIcon />
-                追加
-              </>
-            )}
-          </Button>
-          <DropzoneDialog
-            acceptedFiles={['image/*']}
-            cancelButtonText="cancel"
-            submitButtonText="submit"
-            maxFileSize={500000}
-            open={open}
-            onClose={handleClose}
-            onSave={handleSave}
-            filesLimit={1}
-            previewGridProps={{
-              container: { justify: 'center' },
-              item: { xs: 8 }
-            }}
-            showAlerts={['error']}
-            dialogTitle={label}
-            previewText=""
-          />
-        </Grid>
-      </>
-    </Labeled>
-  )
 }
 
 const EditSessionForm: FC<Omit<SimpleFormProps, 'children'>> = (props) => {
@@ -210,11 +77,17 @@ const EditSessionForm: FC<Omit<SimpleFormProps, 'children'>> = (props) => {
                   validate={[required(), colorValidator]}
                   label="プログレスバー"
                 />
-                <ImageInput
-                  source="images.logo.key"
-                  resource="sessions"
-                  label="ヘッダーロゴ"
-                />
+
+                <FormDataConsumer>
+                  {({ formData }) => (
+                    <ImageInput
+                      source="images.logo.key"
+                      resource="sessions"
+                      label="ヘッダーロゴ"
+                      sessionId={formData.id}
+                    />
+                  )}
+                </FormDataConsumer>
                 <RadioButtonGroupInput
                   source="images.agent"
                   resource="sessions"
