+++ components/Admin/Session/parts/index.ts
@@ -1,0 +1,2 @@
+export { default as ColorInput } from './ColorInput'
+export { default as ImageInput } from './ImageInput'
