+++ components/Chat/Message/Content/String.tsx
@@ -1,6 +1,6 @@
 import { FC, useEffect } from 'react'
 import { StringType } from '@botui/types'
-import { useMessageContext } from '../../../../../hooks/use-message-context'
+import { useMessageContext } from '../../../../hooks/use-message-context'
 
 const String: FC = () => {
   const { message, handleUpdate } = useMessageContext()
